Restructure files in user-service, admin-service and gateway
