--- services/admin-service/openapi-doc.json
@@ -1,67 +1,0 @@
-{
-  "openapi": "3.1.0",
-  "info": {
-    "title": "Admin Service API",
-    "description": "API for granting/revoking administrative roles on app users",
-    "version": "1.0.0"
-  },
-  "servers": [
-    {
-      "url": "http://localhost:4000/"
-    }
-  ],
-  "paths": {
-    "/setAdmin/{githubId}": {
-      "put": {
-        "description": "",
-        "parameters": [
-          {
-            "name": "githubId",
-            "in": "path",
-            "required": true,
-            "schema": {
-              "type": "string"
-            }
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "OK"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    },
-    "/removeAdmin/{githubId}": {
-      "put": {
-        "description": "",
-        "parameters": [
-          {
-            "name": "githubId",
-            "in": "path",
-            "required": true,
-            "schema": {
-              "type": "string"
-            }
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "OK"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    }
-  }
-}
