+++ services/gateway/package.json
@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "private": true,
   "description": "Gateway Service between frontend and backend",
-  "main": "app.ts",
+  "main": "src/app.ts",
   "scripts": {
-    "start": "ts-node-dev ./app.ts",
+    "start": "ts-node-dev src/app.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
