+++ services/gateway/src/auth/firebase.ts
