+++ services/gateway/src/logging/logging.ts
