+++ services/user-service/openapiDocGen.ts
@@ -10,7 +10,7 @@ const doc = {
 };
 
 const outputFile = './openapiDoc.json';
-const endpointsFiles = ['./app.ts'];
+const endpointsFiles = ['./src/app.ts'];
 
 /* NOTE: if you use the express Router, you must pass in the
    'endpointsFiles' only the root file where the route starts,
