+++ services/user-service/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "yarn swagger-autogen && ts-node-dev ./app.ts",
+    "start": "yarn swagger-autogen && ts-node-dev src/app.ts",
     "swagger-autogen": "ts-node-dev ./openapiDocGen.ts",
     "build": "tsc",
     "test": "vitest"
