+++ services/user-service/src/db/functions.ts
